+++ src/pages/DirectMessage/components/MessageList/MessageInput/MessageInput.tsx
@@ -1,5 +1,5 @@
 import * as S from "./MessageInput.Styles"
-import MessageProfile from "../../MessageGroupList/MessageProfile"
+import MessageProfile from "../../MessageGroupList/MessageGroupUserlist/MessageProfile/MessageProfile"
 import SendIcon from "@mui/icons-material/Send"
 import { DM_PLACEHOLDER_MESSAGE } from "@/constants/placeholderMessage"
 import useTextArea from "@/pages/DirectMessage/hooks/useTextArea"
